+++ sbt-scalajs-bundler/src/main/scala/scalajsbundler/sbtplugin/NpmUpdateTasks.scala
@@ -5,13 +5,15 @@ import scalajsbundler.ExternalCommand
 import sbt._
 
 object NpmUpdateTasks {
+
   /**
-    * Runs the Npm or Yarn
-    * @param targetDir npm Directory
+    * Runs either `npm install` or `yarn install` and installs JavaScript resources as node packages.
+    *
+    * @param targetDir npm directory
     * @param packageJsonFile Json file containing NPM dependencies
     * @param useYarn Whether to use yarn or npm
-    * @param jsResources A sequence of javascript resources
-    * @param stream A sbt TaskStream
+    * @param jsResources A sequence of JavaScript resources
+    * @param streams A sbt TaskStream
     * @param npmExtraArgs Additional arguments to pass to npm
     * @param yarnExtraArgs Additional arguments to pass to yarn
     * @return The written npm directory
@@ -23,28 +25,69 @@ object NpmUpdateTasks {
                 jsResources: Seq[VirtualJSFile with RelativeVirtualFile],
                 streams: Keys.TaskStreams,
                 npmExtraArgs: Seq[String],
-                yarnExtraArgs: Seq[String]
-                ) = {
-    val log = streams.log
+                yarnExtraArgs: Seq[String]): File = {
+    val dir = npmInstallDependencies(baseDir, targetDir, packageJsonFile, useYarn, streams, npmExtraArgs, yarnExtraArgs)
+    npmInstallJSResources(targetDir, jsResources, streams)
+    dir
+  }
 
+  /**
+    * Runs either `npm install` or `yarn install`.
+    *
+    * @param targetDir npm directory
+    * @param packageJsonFile Json file containing NPM dependencies
+    * @param useYarn Whether to use yarn or npm
+    * @param streams A sbt TaskStream
+    * @param npmExtraArgs Additional arguments to pass to npm
+    * @param yarnExtraArgs Additional arguments to pass to yarn
+    * @return The written npm directory
+    */
+  def npmInstallDependencies(baseDir: File,
+                             targetDir: File,
+                             packageJsonFile: File,
+                             useYarn: Boolean,
+                             streams: Keys.TaskStreams,
+                             npmExtraArgs: Seq[String],
+                             yarnExtraArgs: Seq[String]): File = {
+    val log = streams.log
     val cachedActionFunction =
       FileFunction.cached(
-        streams.cacheDirectory / "scalajsbundler-npm-update",
+        streams.cacheDirectory / "scalajsbundler-npm-install",
         inStyle = FilesInfo.hash
       ) { _ =>
         log.info("Updating NPM dependencies")
         ExternalCommand.install(baseDir, targetDir, useYarn, log, npmExtraArgs, yarnExtraArgs)
-        jsResources.foreach { resource =>
-          IO.write(targetDir / resource.relativePath, resource.content)
-        }
         Set.empty
       }
-
-    cachedActionFunction(Set(packageJsonFile) ++
-      jsResources.collect { case f: FileVirtualJSFile =>
-        f.file
-      }.to[Set])
-
+    cachedActionFunction(Set(packageJsonFile))
     targetDir
+  }
+
+  /**
+    * Installs JavaScript resources as node packages.
+    *
+    * @param targetDir npm directory
+    * @param jsResources The JavaScript resources
+    * @param streams A sbt TaskStream
+    * @return The paths to the node packages
+    */
+  def npmInstallJSResources(targetDir: File,
+                            jsResources: Seq[VirtualJSFile with RelativeVirtualFile],
+                            streams: Keys.TaskStreams): Seq[File] = {
+    val jsFileResources = jsResources.collect {
+      case file: FileVirtualJSFile => file
+    }.toSet
+    val cachedActionFunction = FileFunction.cached(
+        streams.cacheDirectory / "scalajsbundler-npm-install-resources",
+        inStyle = FilesInfo.hash
+      ) { _ =>
+        jsFileResources.map { resource =>
+          val resourcePath = targetDir / resource.relativePath
+          IO.write(resourcePath, resource.content)
+          resourcePath
+        }
+      }
+    val files = jsFileResources.map(_.file)
+    cachedActionFunction(files).toSeq
   }
 }